app/requirements.txt
gunicorn
django
web3
djangorestframework
markdown
django-filter
python-decouple
django-cors-headers
dj-database-url
psycopg2
django-heroku
whitenoise
django-background-tasks-updated
